CodeEditor uses EditorRoot
