+++ app/gui/src/project-view/components/EditorRoot.vue
@@ -26,23 +26,4 @@ defineExpose({ rootElement })
   width: 100%;
   height: 100%;
 }
-
-:deep(.cm-content) {
-  font-family: var(--font-sans);
-}
-
-:deep(.cm-scroller) {
-  /* Prevent touchpad back gesture, which can be triggered while panning. */
-  overscroll-behavior: none;
-}
-
-.EditorRoot :deep(.cm-editor) {
-  position: relative;
-  width: 100%;
-  height: 100%;
-  opacity: 1;
-  color: black;
-  font-size: 12px;
-  outline: none;
-}
 </style>
