+++ app/gui/src/project-view/components/MarkdownEditor/MarkdownEditorImpl.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import EditorRoot from '@/components/MarkdownEditor/EditorRoot.vue'
+import EditorRoot from '@/components/EditorRoot.vue'
 import { highlightStyle } from '@/components/MarkdownEditor/highlight'
 import {
   provideDocumentationImageUrlTransformer,
@@ -63,6 +63,27 @@ const editing = ref(false)
   />
   <VueComponentHost ref="vueHost" />
 </template>
+
+<style scoped>
+:deep(.cm-content) {
+  font-family: var(--font-sans);
+}
+
+:deep(.cm-scroller) {
+  /* Prevent touchpad back gesture, which can be triggered while panning. */
+  overscroll-behavior: none;
+}
+
+.EditorRoot :deep(.cm-editor) {
+  position: relative;
+  width: 100%;
+  height: 100%;
+  opacity: 1;
+  color: black;
+  font-size: 12px;
+  outline: none;
+}
+</style>
 
 <!--suppress CssUnusedSymbol -->
 <style module>
